Move all Wasm related code into ws_js.go
This way we don't pollute the directory tree.
